% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{Abbrev}
\alias{Abbrev}
\title{Tranforms long numbers into short strings.}
\usage{
Abbrev(value)
}
\arguments{
\item{value}{A single numeric value}
}
\value{
A string value such as 1k for 1000
}
\description{
Abbreviates long numeric values into visually shorter strings
}
\examples{
Abbrev(1700000)
}

