% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{badCDF}
\alias{badCDF}
\title{Make a straight-line faux CDF.}
\usage{
badCDF(range, gran, ...)
}
\arguments{
\item{range}{A vector length 2 containing user-specified min and max to
truncate the universe to}

\item{gran}{The smallest unit of measurement in the data (one [year]
for a list of ages)}

\item{...}{Optionally add additional parameters. This is primarily used to allow automated
execution of varied diagnostic functions.}
}
\value{
A fake CDF for demonstration only.
}
\description{
Creates a placeholder CDF (a uniform straight line) 
   for demonatration.
}
\examples{
badCDF(c(1,50), 1)
}

