% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{diffat75}
\alias{diffat75}
\title{Determine the distance between CDFs at the .75 quantile.}
\usage{
diffat75(Y, est, ...)
}
\arguments{
\item{Y}{The vector output of a non-differentially private 
CDF computation (cumulative count bins)}

\item{est}{The vector output of a differentially private
CDF computation (cumulative count bins)}

\item{...}{Optionally add additional parameters. This is primarily used to allow automated
execution of varied diagnostic functions.}
}
\value{
The error at the .75 quantile
}
\description{
Find the error (between 0 and 1) introduced by DP-Noise 
  at the .75 quantile.
}
\examples{
diffat75(c(.1,.2,.3,.4,.5,.6,.7,.8,.9,1), c(.1,.2,.3,.3,.3,.3,.3,.3,.4,1))
}

