\name{dpCDFtesting-package}
\alias{dpCDFtesting-package}
\alias{dpCDFtesting}
\docType{package}
\title{
  Comprehensively evaluate and visualize the output of dpCDF-generating algorithm implementations. dp = Differential Privacy
}
\description{
  This package's primary contribution is the function CDFtest, which is used to visualize and collect large empirical diagnostic data on the performance of user-defined dpCDF implementations. It also includes 4 simple dpCDF implementations.
  
}
\details{
Use ?CDFtest for best information. Other valuable functions are "functionH", "functionHmono", "functionS2", and "functionSUB", which generate dpCDFs through different methods.
}
\author{
Daniel Muise, Harvard SEAS Privacy Tools group
Kobbi Nissim, Harvard CRCS Privacy Tools group
Georgios Kellaris, Harvard CRCS Privacy tools group

Maintainer: Daniel Muise <dmuise@standord.edu>
}
\references{
See http://privacytools.seas.harvard.edu/
}
\keyword{Differential Privacy}