% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{horzdiffatQuantile}
\alias{horzdiffatQuantile}
\title{Determine the distance between the quantile values 
  returned by two CDFs.}
\usage{
horzdiffatQuantile(Y, est, range, gran, quantile, ...)
}
\arguments{
\item{Y}{The vector output of a non-differentially private CDF
computation (cumulative count bins)}

\item{est}{The vector output of a differentially private CDF 
computation (cumulative count bins)}

\item{range}{A vector length 2 containing user-specified min and max 
to truncate the universe to}

\item{gran}{The smallest unit of measurement in the data (one [year]
for a list of ages)}

\item{quantile}{A quantile value between 0 and 1, defaults to
0.5 for the median}

\item{...}{Optionally add additional parameters. This is primarily used to allow automated
execution of varied diagnostic functions.}
}
\value{
The horizontal error at the quantile specified by \code{quantile}
}
\description{
Find the distance between the quantile value and that returned
  by the dpCDF at a given quantile.
}
\examples{
diffatQuantile(c(.1,.2,.3,.4,.5,.6,.7,.8,.9,1), 
   c(.1,.2,.3,.3,.3,.3,.3,.3,.4,1),c(1,10), 1, .05)
}

