% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{KurtDiffpdf}
\alias{KurtDiffpdf}
\title{Error in Kurtosis from CDF (under development)}
\usage{
KurtDiffpdf(Y, est, gran, range)
}
\arguments{
\item{Y}{The vector output of a non-differentially private CDF
computation (cumulative count bins)}

\item{est}{The vector output of a differentially private CDF 
computation (cumulative count bins)}

\item{gran}{The smallest unit of measurement in the data (one [year]
for a list of ages)}

\item{range}{A vector length 2 containing user-specified min and max
Note that the gran and range must be the same as used to make the DP-CDF!}

\item{...}{Optionally add additional parameters. This is primarily used to allow automated
execution of varied diagnostic functions.}
}
\value{
A single difference value
}
\description{
Calculate difference between the private Kurtosis and the original Kurtosis (from CDFs)
}

