% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{L2empiric}
\alias{L2empiric}
\title{Calculate the empirical L2norm between two CDFs.}
\usage{
L2empiric(Y, est, ...)
}
\arguments{
\item{Y}{The vector output of a non-differentially private CDF 
computation (cumulative count bins)}

\item{est}{The vector output of a differentially private CDF
computation (cumulative count bins)}

\item{...}{Optionally add additional parameters. This is primarily used to allow automated
execution of varied diagnostic functions.}
}
\value{
The empirical L2 norm
}
\description{
Calculates the L2 (squared error) area between the 
   non-private CDF and the dpCDF
}
\examples{
L2empiric(c(.1,.2,.3,.4,.5,.6,.7,.8,.9,1), c(.1,.2,.3,.3,.3,.3,.3,.3,.4,1))
}

