% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{MaxError_PDF}
\alias{MaxError_PDF}
\title{Determine an approximate PDF's maximum error.}
\usage{
MaxError_PDF(Y, est, ...)
}
\arguments{
\item{Y}{The vector output of a non-differentially private PDF 
computation (heights of bins)}

\item{est}{The vector output of a differentially private PDF
computation (heights of bins)}

\item{...}{Optionally add additional parameters. This is primarily used to allow automated
execution of varied diagnostic functions.}
}
\value{
A single value, the largest absolute vertical difference between 
   parallel observations in the private- and true-PDF vectors.
}
\description{
Find the maximum direct error between a non-private PDF and a 
   DP approximation of that PDF.
}
\examples{
MaxError_PDF(c(.1,.2,.3,.4,.5,.6,.7,.8,.9,1), c(.1,.2,.3,.3,.3,.3,.3,.3,.4,1))
}

