% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{MovetoRange}
\alias{MovetoRange}
\title{Clamp a value to a specified range.}
\usage{
MovetoRange(val, range)
}
\arguments{
\item{val}{A value to clamp.}

\item{range}{A vector of length 2 in the form \code{c(min, max)}}
}
\value{
A single value that is either unchanged or clamped upward to minimum 
   or clamped downward to the maximum
}
\description{
Returns a vector of elements clamped to the specified minimum
   and maximum
}
\examples{
MovetoRange(11, c(1,10))
}

