% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{MSE}
\alias{MSE}
\title{Calculate the MSE of a DP-CDF relative to the non-private CDF.}
\usage{
MSE(Y, est, ...)
}
\arguments{
\item{Y}{The vector output of a non-differentially private CDF
computation (cumulative count bins)}

\item{est}{The vector output of a differentially private CDF
computation (cumulative count bins)}

\item{...}{Optionally add additional parameters. This is primarily used to allow automated
execution of varied diagnostic functions.}
}
\value{
The MSE
}
\description{
Calculates the Mean Squared Error area between the non-private 
    CDF and the DP-CDF
}
\examples{
MSE(c(.1,.2,.3,.4,.5,.6,.7,.8,.9,1), c(.1,.2,.3,.3,.3,.3,.3,.3,.4,1))
}

