% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{nodes}
\alias{nodes}
\title{Node parser.}
\usage{
nodes(height, k, l)
}
\arguments{
\item{height}{The height of the tree}

\item{k}{The tree degree}

\item{l}{The leaf length}
}
\value{
A nodesum containing information for MSEanalytic
}
\description{
Runs through tree nodes (assists MSE analytic)
}
\examples{
nodes(10,4,2)
}

