% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Smooth}
\alias{Smooth}
\title{Monotonicity enforcement}
\usage{
Smooth(x)
}
\arguments{
\item{x}{A numeric vector to be enforced}
}
\value{
A monotonized vector
}
\description{
When CDFs get out of line, we call the enforcer
}

