% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFtestingSuite.R
\name{smoothVector2}
\alias{smoothVector2}
\title{Enforce monotnocity on a vector.}
\usage{
smoothVector2(cdf)
}
\arguments{
\item{cdf}{The vector output of a differentially private CDF 
computation (cumulative count bins)}

\item{...}{Optionally add additional parameters. This is primarily used to allow automated
execution of varied diagnostic functions.}
}
\value{
A single monotonically increasing vector which is the post-processed
   DP-CDF's Y coordinates
}
\description{
Forces DP-CDFs into the nearest monotonic vector
   (by euclidean distance minimization).
}
\examples{
smoothVector2(c(.1,.2,.3,.2,.3,.3,.3,.3,1))
}

