% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TreeCDF}
\alias{TreeCDF}
\title{Creates a Tree then a CDF}
\usage{
TreeCDF(eps, ds, Ks, methods, mins, maxs, grans, datas)
}
\arguments{
\item{eps}{An epsilon value for Differential Privacy}

\item{ds}{The data or something}

\item{Ks}{the degree of the tree}

\item{methods}{Either H or S2 or SUB}

\item{mins}{the minimum of the domain's range}

\item{maxs}{the maximum of the domain's range}

\item{grans}{The granularity}

\item{datas}{The data to be CDFd}
}
\value{
A dpCDF
}
\description{
This thing sure does make a fine CDF
}

